Color edges according to source node.
